+++ lib/react/rails/component_mount.rb
@@ -56,7 +56,10 @@ module React
         unless prerender_options == :static
           html_options[:data].tap do |data|
             data[:react_class] = name
-            data[:react_props] = (props.is_a?(String) ? props : props.to_json)
+            data[:react_props] = props_to_json(
+              props,
+              null_to_undefined: Dummy::Application.config.react.null_to_undefined_props
+            )
             data[:hydrate] = "t" if prerender_options
 
             num_components = @cache_ids.count { |c| c.start_with? name }
@@ -65,6 +68,17 @@ module React
         end
 
         html_options
+      end
+
+      def props_to_json(props, options = { null_to_undefined: false })
+        return props if props.is_a?(String)
+        return props.to_json unless options[:null_to_undefined]
+
+        # This regex matches key:value with null values while ensuing no string with similar
+        # pattern gets matched. It doesn't include null values in arrays.
+        props.to_json
+             .gsub(/([^\\]":)null([,}\]])/, '\1undefined\2') # match simple null values
+             .gsub(/([^\\]":(\[[^\\"]+,|\[))null([,\]])/, '\1undefined\3') # Match nulls in array
       end
 
       def rendered_tag(html_options, &block)